# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Username or email already exists"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Register a new user account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Returns access and refresh JWT tokens on success.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT tokens"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Login with username/email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Submit the refresh token in the Authorization header to obtain a new access token.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/contests/{number}": {
      "get": {
        "operationId": "ContestsController_detail",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Contest number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest detail"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "summary": "Get contest detail",
        "tags": [
          "contests"
        ]
      }
    },
...
    "/admin/judge-nodes": {
      "get": {
        "operationId": "AdminController_judgeNodes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of judge nodes with heartbeat info"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all judge nodes and their status",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/config": {
      "get": {
        "operationId": "AdminController_getConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of config key-value pairs"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all system configuration entries",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_setConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string"
                  },
                  "value": {
                    "type": "string"
                  }
                },
                "required": [
                  "key",
                  "value"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config entry created or updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a system configuration entry",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/rejudge/candidates": {
      "post": {
        "operationId": "RejudgeController_listCandidates",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejudgeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of submissions matching the rejudge criteria"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List rejudge candidates",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/rejudge/preview": {
      "post": {
        "operationId": "RejudgeController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejudgeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preview of affected submissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview rejudge impact",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/rejudge": {
      "post": {
        "description": "Rejudge all submissions matching the criteria. This is an async operation.",
        "operationId": "RejudgeController_execute",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejudgeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejudge initiated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute rejudge",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contests": {
      "get": {
        "operationId": "AdminContestsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all contests including hidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all contests (admin)",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminContestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contest created"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new contest",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contests/{id}": {
      "get": {
        "operationId": "AdminContestsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest detail"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get contest detail by ID (admin)",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminContestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest updated"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a contest",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminContestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contest deleted"
          },
          "404": {
            "description": "Contest not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a contest",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contests/{id}/registrations": {
      "get": {
        "operationId": "AdminContestsController_listRegistrations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of registrations"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List contest registrations",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminContestsController_addRegistration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpsertContestRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created or updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add or update a contest registration",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contests/{id}/registrations/{userId}": {
      "patch": {
        "operationId": "AdminContestsController_updateRegistration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Contest ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateContestRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a contest registration",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "EpochJudge API",
    "description": "纪元 — 开源 Online Judge 平台 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "RejudgeRequestDto": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "problem",
              "submission",
              "contest"
            ]
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "submissionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "statuses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "scope",
          "ids"
        ]
      },
      "CreateContestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "judgeMode": {
            "type": "string",
            "enum": [
              "OI",
              "ACM"
            ]
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          },
          "freezeAt": {
            "type": "object"
          },
          "accessPassword": {
            "type": "object",
            "description": "明文比赛密码，空表示无密码"
          },
          "problemIds": {
            "description": "题目 ID 列表，顺序即题号 A/B/C…",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "visibility",
          "judgeMode",
          "startAt",
          "endAt"
        ]
      },
      "UpdateContestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "judgeMode": {
            "type": "string",
            "enum": [
              "OI",
              "ACM"
            ]
          },
          "startAt": {
            "type": "string"
          },
          "endAt": {
            "type": "string"
          },
          "freezeAt": {
            "type": "object",
            "nullable": true
          },
          "accessPassword": {
            "type": "object",
            "nullable": true
          },
          "problemIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminUpsertContestRegistrationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "isStarTeam": {
            "type": "boolean"
          }
        },
        "required": [
          "userId"
        ]
      },
      "AdminUpdateContestRegistrationDto": {
        "type": "object",
        "properties": {
          "isStarTeam": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}